Update P4BridgeServer.cpp for crash fix
The p4bridge expriences a crash when the application name is over 16 characters long. This is because pProgramName has been changed from a char * to a std::string. The std::string is only 16 bytes in length. So any application that we have that's more than 16 characters in length is memory stomping the string.